+++ nym-registration-client/src/clients/lp/dvpn.rs
@@ -7,8 +7,8 @@
 //! client borrows a channel for as long as it takes to register over it, which keeps the two apart:
 //! a data-plane client wants [`LpGatewayClient`] with none of this attached.
 
-use crate::lp_client::bandwidth_claim::produce_bandwidth_claim;
-use crate::lp_client::helpers::{LpFrameDeliverExt, LpFrameSendExt};
+use crate::clients::lp::bandwidth_claim::produce_bandwidth_claim;
+use crate::clients::lp::helpers::{LpFrameDeliverExt, LpFrameSendExt};
 use nym_bandwidth_controller::BandwidthTicketProvider;
 use nym_credentials_interface::TicketType;
 use nym_crypto::asymmetric::{ed25519, x25519};
